refactor: change type to get sync with backend
