+++ apps/frontend/src/app/types/blog-post.type.ts
@@ -1,0 +1,10 @@
+export type BlogPostType = {
+    id: string
+    title: string
+    description: string
+    content: string
+    tags: string[]
+    author: string
+    createdAt: Date
+    updatedAt: Date
+}
