--- apps/frontend/src/app/types/post.type.ts
@@ -1,7 +1,0 @@
-export type PostType = {
-    id: number
-    title: string
-    description: string
-    date: Date
-    tags: string[]
-}
